Factor out `StorySelector` into a separate module
